+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * typecheck `string->symbol` and `symbol->string`
 * fix `parent.frame` inside `Intepreter`
 * fix `eval` without env
+* fix quote as delimiter
+* fix comma dot as two tokens
 
 ## 1.0.0-beta.14
 ### Breaking